+++ nodes/__init__.py
@@ -1,7 +1,9 @@
 from .image_comparer import DossImageComparer, choose_comparison_images
+from .save_image import DossSaveImage
 
 
 __all__ = [
     "DossImageComparer",
+    "DossSaveImage",
     "choose_comparison_images",
 ]